+++ test/fixtures/CHANGELOG.md
@@ -7,8 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 10.1.0-beta1 - 2021/02/08
 
-- Use modern workflow to package and release the plugin
-- Minor code changes
+- This is the latest documented version in this changelog
+- The changelog module is tested against these lines
+- Be careful modifying this file
 
 ## 10.1.0-alpha1 - 2021/01/10